Struct google_doubleclickbidmanager1::ReportListreportCall

pub struct ReportListreportCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Retrieves stored reports.

A builder for the listreports method supported by a report resource. It is not used directly, but through a ReportMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.reports().listreports("queryId")
             .doit();

Methods

impl<'a, C, A> ReportListreportCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, ListReportsResponse)>

Perform the operation you have build so far.

fn query_id(self, new_value: &str) -> ReportListreportCall<'a, C, A>

Query ID with which the reports are associated.

Sets the query id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn delegate(self, new_value: &'a mut Delegate) -> ReportListreportCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> ReportListreportCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.