Struct google_doubleclickbidmanager1::LineitemMethods

pub struct LineitemMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on lineitem resources. It is not used directly, but through the DoubleClickBidManager hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_doubleclickbidmanager1 as doubleclickbidmanager1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use doubleclickbidmanager1::DoubleClickBidManager;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = DoubleClickBidManager::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `downloadlineitems(...)` and `uploadlineitems(...)`
// to build up your call.
let rb = hub.lineitems();

Methods

impl<'a, C, A> LineitemMethods<'a, C, A>

fn uploadlineitems(&self, request: UploadLineItemsRequest) -> LineitemUploadlineitemCall<'a, C, A>

Create a builder to help you perform the following task:

Uploads line items in CSV format.

Arguments

request - No description provided.

fn downloadlineitems(&self, request: DownloadLineItemsRequest) -> LineitemDownloadlineitemCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves line items in CSV format.

Arguments

request - No description provided.