Struct google_doubleclickbidmanager1::QueryMetadata

pub struct QueryMetadata {
    pub google_cloud_storage_path_for_latest_report: Option<String>,
    pub data_range: Option<String>,
    pub send_notification: Option<bool>,
    pub locale: Option<String>,
    pub google_drive_path_for_latest_report: Option<String>,
    pub format: Option<String>,
    pub share_email_address: Option<Vec<String>>,
    pub report_count: Option<i32>,
    pub running: Option<bool>,
    pub latest_report_run_time_ms: Option<String>,
    pub title: Option<String>,
}

Query metadata.

This type is not used in any activity, and only used as part of another schema.

Fields

google_cloud_storage_path_for_latest_report

The path to the location in Google Cloud Storage where the latest report is stored.

data_range

Range of report data.

send_notification

Whether to send an email notification when a report is ready. Default to false.

locale

Locale of the generated reports. Valid values are cs CZECH de GERMAN en ENGLISH es SPANISH fr FRENCH it ITALIAN ja JAPANESE ko KOREAN pl POLISH pt-BR BRAZILIAN_PORTUGUESE ru RUSSIAN tr TURKISH uk UKRAINIAN zh-CN CHINA_CHINESE zh-TW TAIWAN_CHINESE

An locale string not in the list above will generate reports in English.

google_drive_path_for_latest_report

The path in Google Drive for the latest report.

format

Format of the generated report.

share_email_address

List of email addresses which are sent email notifications when the report is finished. Separate from sendNotification.

report_count

Number of reports that have been generated for the query.

running

Whether the latest report is currently running.

latest_report_run_time_ms

The time when the latest report started to run.

title

Query title. It is used to name the reports generated from this query.